backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
pymongo==4.10.1
PyJWT==2.10.1
bcrypt==4.2.1
python-dotenv==1.0.1
pydantic[email]==2.10.3
python-multipart==0.0.9
certifi==2025.1.31
rapidfuzz==3.13.0
